Organize components in separate folders
